better
there is still a jump when the first measure of a line is edited.  this is because it is usually the highest (low y) and so moves down.  accum. needs to keep track of how much max measure (lowest) moves, not min.  min could move a lot and max not at all.